requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
blinker==1.9.0
click==8.3.1
discord.py==2.6.4
Flask==3.1.2
frozenlist==1.8.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.0
propcache==0.4.1
python-dotenv==1.2.1
typing_extensions==4.15.0
Werkzeug==3.1.4
yarl==1.22.0
requests
python-dateutil
